make SNMP fetcher paths relative
For the relay we want to be able to create the
fetchers with different base paths.

This currently only covers SNMP fetchers.
At some point we will have to cover other fetchers
as well, but currently it is not clear how we even
deal with other paths (like the UUID link file
or executable paths).

Change-Id: I5bf052a98fbc6c570cc6ea998a339a0dbc56efdc