18529 OTel: Avoid opening TCP port when sending logs to Event Console
Previously when the option "Send log messages to event console" was enabled in the collector configuration, the communication happened via TCP.
This meant port 514 had to be opened, although the communication is local.
With this change, the communication occurs over a local socket, meaning no port has to be opened.

CMK-25004

Change-Id: If850e995b0d0ac5392bbc6d2740305c3aaec805e