Azure SQL DBs: Allow setting connection thresholds
Currently, no ruleset is given for Azure SQL Databases, so there is
simply no way to set these levels as an end-user.

Create a ruleset for this and make use of it.

Note that this cannot use the existing (legacy, WATO) ruleset for azure
database connections, because Azure SQL provides
"successful_connections" and "failed_connections, rather than
"active_connections" and "failed_connections". So this ruleset gets to
be its own unique snowflake.

Change-Id: I623711dc19addcd06d1f2af4472e708a8a87bb97