Werk tool: Make 1st `werk new` less confusing
If you run `werk new` without having reserved any IDs before, then the
project key simply doesn't exist in the stash file and we only showed a
KeyError traceback in this case with no explanation about how to fix
it.

Catch the KeyError and tell the user what to do.

Change-Id: Idffdd43e0e808c22a03a735b09886b08ad59089a