Enable global settings to affect more than one configuration domain
This is needed for a new global setting that will configure the ClickHouse
instance used by a Checkmk site. The primary configuration domain of this
setting will be `ConfigDomainGUI`, since it will be needed in various places in
`cmk.gui`. However, changing this setting must additionally trigger
`ConfigDomainOTelCollector.activate`, because the ClickHouse connection
parameters will be part of the OTel collector configuration in the future.

CMK-26124

Change-Id: Ic2a5a027a2dbe53dcd7737a8e5de1fe5154bbc06