test-plugin-update: Update base version
The base version is here changed from 2.4.0p3 to 2.4.0p12.

Reason: we are now testing Werk #18568 too, whose fix is included
starting from cmk v2.4.0p12. If we want to avoid a crash report to be
generated during the test's execution, such base version (or higher)
needs to be used.

Change-Id: I06098e10337af79b8a901045b658ca03098c4615