Implement mechanism to add affected configuration domain to global setting
Use case: We will implement a global setting to configure the ClickHouse
instance used by a Checkmk site. `ConfigDomainOTelCollector` is affected by
changes to this variable, since the ClickHouse instance is part of the OTel
collector configuration. However, the primary config domain of this new global
setting will be `ConfigDomainGUI`. Furthermore, we don't want to import
`ConfigDomainOTelCollector` in the place where we define and register this new
variable. Rather, we want the dependency to be the other way round: The OTel
collector code should import the new config variable and add
`ConfigDomainOTelCollector` as an affected config domain.

CMK-26124

Change-Id: Id7377693c1270a44e521b4fbe842781ac23719d3