18475 FIX db2_logsizes: fix crash when write is suspended
`SELECT total_log_used from sysibmadm.snapdb` (which is used internally
by db2_logsizes) returns `"-"` when writes are suspended in db2.

The check did not consider this edgecase and crashed with `ValueError:
invalid literal for int() with base 10: '-'`. Now the check will go to
`UNKNOWN` and show `Can not read usedspace`.

SUP-25579

Change-Id: I0233b92f92e9e587ac23208d5cc5f11ff1594cc8