Build relay OCI image using bazel
With the initial release of the relay, the container image will be the
primary way we deliver the relay to our users. It enables the update
concept we are about to build (SAASDEV-5184) in one of the next steps of
the project.

This change lays the foundation for that.

Add rules_oci to bazel and introduce a basic image build process.
Currently the following is supported:

1. Build an OCI image
2. Load image into local podman instance
3. Load image into local docker instance

Details on how to use it can be found in the README.md.

SAASDEV-5436

Change-Id: Ic1bc0be9d66c0aa55f64628e80551fb3c2786c65