Fix internal error with custom snapins
This issue was not but should have been found by the typing system.
Unfortunately the situation around the custom snapins, with the dynamic
creation of classes, is preventing the system from doing it's job in
this area. We will have to rework this in order to make the area less
error prone.

CMK-26013

Change-Id: I0adf883ab9ed6e8a19c7a2acd969606900dfc786