fix shorting of duplicated prefixes
the idea of this part of the code is to shorten unnecessary long selector
names: imagine the file FormSomething.vue in `./src/form` wich would
demand `.form-form-something` as block name, which is then reduced to
`.form-something`.

the previous implementation had problems with shorting if the folder
name started the same as the component name: `./src/forms` (note plural
s!) and a file `FormSomething` was demanding `.form-something` as block
name.

after this fix stylelint will request the selector `.forms-form-something`

Change-Id: I125ac63caeba8ca4c5f40fcdb3531fc21e3462bb