18643 Livestatus Labels Table Has Two New Columns
The Livestatus labels table is used to query the `name` and `value` of labels for all entities on a site.

With this change, two new columns are added - `type` and `source`.

The `type` column refers to the type of entity which the label belongs to, either `service` or `host`.

The `source` column refers to how the label was created, either `discovered`, `ruleset` or `explicit`.

Any queries which do not have column constraints must be updated to constrain the column names.

If not that, then the corresponding logic should be updated to handle the new columns.

If your queries have column constraints already then no changes are required.

Change-Id: I7fd881807f2e9d7a978d328790f4b3336aafb63e