ch: add missing user.xml
ClickHouse dispatches different c++ functions depending on the binary
name. If we use `clickhouse-server`, then this configuration is
necessary, because otherwise an error is report about the missing
profile. For now, these are placeholder values. `clickhouse-server` will
be useful in our init scripts.

We also add `http_port` and `tcp_port` for the same reason.

CMK-25860

Change-Id: I704735ca664350a09e75bfb5049fbacc7d88c2f9