Use Checkmk language settings inside vue
Retrieve current language from lang attribute from the document and only
render vue apps once translation is loaded.

This is necessary since single page language switching is only supported
for reactive variables and some of our code already does things like

`const foo = [_t("bar")]`

Which means the correct language needs to be rendered on the first
pass. This is a general limitation of javascript.

To get around this, we immediately lazy load the requested language and
postpone rendering until the requested language is available.

If we were to render english first, the non-reactive texts wouldn't
update even after the language is loaded.

This behavior is analogous to how rendering would be handled if
languages were loaded synchronously (rendering would block until
everything is in memory).

CMK-24761

Change-Id: I8b086f00ec25a92fbdfcac85f3bd6455eefa4e00