Pass user_permissions to LoggedInUser.may
The core of this change is to pass it to the LoggedInUser.may method, so
that it can make use of it instead of creating a new UserPermissions
object on every call.

For this, it needs to be passed through the user context, which in turn
triggers a huge amount of follow-up changes to make it work.

The UserPermissionSerializableConfig had to be introduced to pass the
needed configuration through background job calls, which needs a
serializable object. All background jobs calling
job_interface.gui_context now need to have this object passed.

This escalated a bit, but most of it is mechanic.

Change-Id: I94141f2a37526e5a28108822b442cfa38e3d3e7a