Add resource attribute used for host name to metrics association host attribute
Conceptually, this field is unnecessary, since this information can be encoded
as an additional attribute filter. However, the current DCD API does not really
allow for setting dynamic host attributes. We would need the DCD to set the
attribute to a different value for each host. Specifically, the filter encoding
the resource attribute used to set the host name would need to be different for
each host. The current API does not really allows this.

Some details:
* Ideally, we would create one `HostOrder` object per host. We could then have
  different attributes for each created host.
* However, with the current API, we create one `HostOrder` object per host
  creation entry. The backend later on links the created hosts back to the
  `HostOrder` objects.

Change-Id: I302b2e5b20d3f918aa5c6e3546ea470a1d1e6fea