Add additional validation for parameter
We used to validate further down the call tree. We should keep the
validation in the lower areas since it is also used in the automation
call. Having it in the page makes it more obvious. The runtime costs are
negligible.
Also added some doctest to clear up what validation that function does.

CMK-26027

Change-Id: I705269e0bd0e608046b700d9d0287f1cf4e3cd8e