Single jobs: receive artifacts from trigger-cmk-distro-package by default
Without this change artifacts are retrieved from
sign-cmk-distro-package. This leads to blocking more resource than
required for jobs that are depending on having a new package built,
because the job for signing a package will request resources while
waiting for the initial package built.
To reduce the usage we switch to the job trigger-cmk-distro-package
which has been optimized to use fewer resources for package builds.

References CMK-26356

Change-Id: I97edff6a6c41759c15fe7e060564fd606245e8e0