skip test_bom_csv_synchronous without date limit
The time component has lead to not having nightly builds around, due to
all package tests failing.
We can't predict a good time for this to happen next, even during the
day this could affect our ability to make releases.
In the interest of the person watching over our builds the next time
this comes around, the time limit has been removed.

There is already a follow-up ticket created to unskip this test and I
prefer the person unskipping dealing with the fallout in CV instead
of all who rely on having a nightly build.

References CMK-23389

Change-Id: I93403688e9fb6b2c399b7ed132de9ca5219a4b2c