test(search): add integration tests
These tests try to test the behavior of the search mechanism as it
affects the results. Namely, the more precise a query, the fewer
results that should be returned. We also want to have a form of
regression test to verify the shape of the data that is being returned
from the endpoint as our frontend relies on it.

CMK-26004

Change-Id: Iaed03fcf9eed58bb304328ee17221fe0cfb059a5