18345 FIX net-snmp: Prevent segfault due to null pointer
Under certain conditions, a function in the net-snmp library could return a null pointer instead of a valid one.

When such a null pointer was accessed a segmentation fault occurred.
This caused Checkmk helper processes to crash and produce core dumps.

This werk patches the library to add validation in order to prevent it from accessing null pointers.

SUP-24082

Change-Id: I44fd47d09fb0a762949209a95f85cd1dbc6ddfe0