18099 FIX ALLNET IP Sensoric: Display errors when special agent failed executing
The special agent for the ALLNET IP Sensoric XML API previously
swallowed errors silently when fetching data, causing failures to go
unnoticed. With this change, the special agent now exits with an error code and
displays the corresponding error message as part of the Check_MK
service. As a result, operators will be promptly alerted to connectivity
or API issues and can take corrective action. No configuration changes
are required, but users should monitor the Check_MK service status for
such errors.

Change-Id: I71eb06bb2f4380645bd0751ab1f08ac906f30acc