Get rid of lib_cmk_repo_no_meraki target
When scanning meraki together with the rest of lib_cmk using rules_mypy,
we get an error: "Source file found twice" error, which is related to
the isolation of the site-packages and __init__.py created in the
site-packages by rules_python.

This change works around the issue in a different way. It adds a custom
mypy wrapper that filters the problematic 'meraki' MYPYPATH entry. The
library is not typed anyways, so there is no drawback in terms of typing
coverage.

Change-Id: I645479f17e8a0e93c2befde017b6a80a2bb87a3c