confg: attach http_proxies to loaded config
This change adds the base configuration parameter `http_proxies`
to the `LoadedConfigFragment` object, and replaces all accesses
to the global variable by an access to the passed down attribute
of the loaded config instance.
This is a) in line with our general direction, and b) highlights
the fact that the only use of the global proxy config apart from
the SSC plugins is via `make_http_proxy_getter`.

Change-Id: I3d5a37949b95e534282d20ae2baddcf11b60c3f2