csv bulk import refactor: Ignore space-only lines
As an edge case of enforcing consistent field counts, it is nicer to the
user if we ignore not only blank lines (already done) but also lines
containing only spaces (which would count as a line of field count 1).

If a line only has whitespace, ignore it.

This is mostly just for convenience and making for a slightly nicer user
experience in case some extra space gets copied in or something.

Change-Id: I7b9c0b18abb7a307050a47c706363c14b404a6b6