18958 FIX Host/Rules incorrectly shown as matching
When clicking on Host/Rules in the "Properties of host" page,
rules were beings shown as matching even though they weren't
(e.g. 3/3 instead of 2/3).

Also, when clicking additionally on the individual rule names, rules
that were not matching the host were being marked as "matching" with a
green badge.

This was happening for hosts in different folders as the ones matching.

These issues are now fixed so that only the matching rules are shown
as "matching" in both the Host/Rules (Rule search) page as in the
individual rule pages.

CMK-26266

Change-Id: I0aa89f07b13dda7d2ce00b682b676b26703766b4