18732 FIX cluster_mode: Improvde service summary
When using a clustered host and aggregating services,
it would happen that in a failover mode with a designated primary host,
the wrong host would be shown as active host in the service summary.

This Werk fixes this issue and always shows the primary as active host.

SUP-25533

(cherry picked from commit 2492bad985391b84c4b3e0877abf0555414ca335)
Change-Id: If93639391c1f59778fcb57c42f01cfad883427dc