Get rid of ReportStore
The report store provided cached access to all reports and the permitted
reports. With a better structured approach and passing of the centrally
loaded data structures from the entry points, we can eliminate the need
for the cached access and replace the global store construct with
straight forward function calls.

In a few cases we trigger the uncached loading deeper in the call stack,
but this seems to be acceptable.

Change-Id: I85aa24099966401b97d0ec4b70b4ad97a96791e0