services: move settings check to helper
This makes it more robust if the default parameters aren't passed to
discovery for whatever reason. It also makes the intent more clear that
if the settings are empty, there's nothing to do.

As a bonus, we longer need to pass the default discovery parameters to
the test cases explicitly.

Change-Id: Ib3207b83ae40ecbc84ea020423dfbd7c221f17db