gui(ajax): move global access out of init
In order for us to be able to register pages as instances, it's
important that we move any logic that accesses global variables like
request or session out of `__init__` for our AJAX pages.

By doing this outside the scope of the request, we've created a temporal
dependency that only allows us to initialize the page right before
handling an HTTP request.

CMK-26754

Change-Id: Ic01df6052379ae8f42af97cbd2f13c08dc5254c0