gui(page): move global access out of init
In order for us to be able to register pages as instances, it's
important that we move any logic that accesses global variables like
request or session out of `__init__` for our pages.

By doing this outside the scope of the request, we've created a temporal
dependency that only allows us to initialize the page right before
handling an HTTP request.

We've also move the base `_from_vars` method definition into the `Page`
class so that pages (not only AJAX pages) can make use of this pattern.

CMK-26754

Change-Id: I801d25c3da0d655d109a635b8307180d3c6f8c7e