fix proxy password lookup
For the lookup of the passwords potentially configured
in a global proxy, we do not need the 'merged' passwords
file, it is enough to look at the globally configured
secrets. More importantly: We can not read the 'merged'
passwords file in the autimation that is meant to (re-)
create it.

Change-Id: Ied1772f57c830d1933cdc48c4632fd2ea21b9234