tests: search for unsigned/unchecked files
CMK-26639

mk-oracle.exe needs to be checked for signing as well
(starting with 2.5.0)

In addition to explicitly checking the signing of a given list of files,
we want to ensure that in the future added files are not forgotten.

To make signing checking and status visible in the package builds,
we need to raise the log level to INFO for tests-packaging.

Change-Id: I62a033d025c47f1cac0547e1b92eba4247b1d2eb