Make test-gerrit-single artifacts directly accessable from parent job
See https://tribe29.slack.com/archives/C01EA6ZBG58/p1760693706024659

The artifacts on the test-gerrit job overview table are only accessable
after the archiveArtifacts function of test-gerrit completed. This
function can only be called once on a job and is thereby called at the
very end of the job.
This causes waittimes for the developer if he does not go to the already
completed sub job (test-gerrit-single) and access the file there

Change-Id: I75a46f247a5ac8aee57307ced00a1da537493957