18478 FIX ibm_storage_ts_drive: handle empty string for read/write errors
A customer reported that IBM TS3100 tape library devices return an empty
string instead of the number of recovered write errors. This happens
sporadically and the correct number of recovered write errors is
reported in the next check cycle.

Prior to this werk the plugin was crashing in that situation, now the
check will return state "unknown" and the summary will indicate that it
received an empty string.

SUP-25829

Change-Id: I1e956a9fb882ef35bfab6b807e9124133300bf85