Call "Prepare Workspace" at trigger-cmk-distro-package job
... to build here the windows artifacts required and requested by nested
jobs like "build-cmk-distro-package" and "sign-cmk-distro-package" as
they do call this stage as well and would request a different windows
artifact otherwise.

This is not an issue, just a flaw, causing windows builds twice. Once
here at "trigger-cmk-distro-package" and then at as well on
"build-cmk-distro-package" which is the correct one then. This should
be avoided to keep the runtime of the "build-cmk-distro-package" job as
short as possible to not allocate resources and of course to not build
things twice.

This is getting more and more worse "just" to have cached Windows
artifacts.

Follows also CMK-26121

Change-Id: I09e3f39f56af77df8e0d709f34588730031cd9f6