Allow the for winagt-build-modules to be built without using the cache.
The logic for downloading the python.cab file was previously hard-coded.
There are times when we need to force the rebuild of python for testing
purposes.

This change passes the DISABLE_CACHE CI parameter through the build
scripts such that we can force the rebuild of python by setting
DISABLE_CACHE to true in the CI.

Preparing for CMK-26899

Change-Id: Ie0161a22773ab1a2ae74e5833f71f556b0d61d08