Introduce global settings context
We want to make the `ValueSpec` of a global setting context-sensitive.
Concretely, we want the default values to depend on the ID of the site for
which the setting is being configured. However, we want to avoid accessing
global variables (for example request variables in this case, implementing the
functionality via this approach would be ugly and brittle). Therefore, we
introduce `GlobalSettingsContext`. This object allows for transporting such
context information into the `ValueSpec`s of global settings.

CMK-27002

Change-Id: Ie73375cf6e1e7dac575f3d8c11a93c1f54b5fcd3