fix test test_windows_artifacts_are_signed for cma packages
for cma packages the returned paths seem to be (only) prefixed with "{omd_version}/" instead of "/opt/omd/versions/{omd_version}/"

we want to be able to handle and compare both cases.

note:
Files were signed correctly and signing was verified as well.
This is "only" about a stable comparison of "available exe files" against "signed files".

Change-Id: I12f66a243b1ee8fa61e02e2ca852bfeefa814798