Adapt testlib edition heuristic
Commit 2b2e739d removed the directories that were used to determine the
editions based on the presence of certain directories. As a consequence
the testlib edition heuristic does not work anymore.

We could now use other directories instead, but the previous mechanism
was kind of a lie anyways, because we don't have edition specific
repositories and test execution environments currently. There is only
the internal and the public repository. This commit adapts to this
reality.

Change-Id: I42ea4038b094545c3ffd42fac71f60b427fd856c