msexch_dag: Add proper parse function
This slighly adapts the logic, since it (more correctly than before)
starts parsing the tables with the first entry instead of
"DatabaseName".

This should have no effect in production as long as we are sure
that the relevant entries always come after "DatabaseName".
Otherwise this would have been an error anyways.

Change-Id: I24800ba2dc7fca16613f5ddbac5ccad3998d3fdb