Remove short edition from the package build
The end goal is to completely get rid of the "short" edition identifier.

As this is a quite fundamental change, it needs to be taken step by step.
For example, the short names will still be used in the werks for some
time until we clean them up there. Likewise a few other places will need
a transition phase.

We now lay the foundation by removing the short edition from the package,
which will trigger some immediate follow-up changes in the system level
tests.

CMK-27226

Change-Id: Id974de890f02e8e0fea60480ac3868a642150108