18137 Undo contact group assignment rule renaming
In Werk #18697, the rules "Assignment of hosts to contact groups" /
"Assignment of services to contact groups" were renamed to "Set host
visibility for contact groups"/"Set service visibility for contact
groups" in an effort to make their effects more clear.

The new names however did not achieve what they set out to do and the
renaming is reverted with this Werk.

Change-Id: Iac3e65dade1dc26af39e3a559563ddde5898f872