CIPARAM_PATH_HASH is set for build-linux-agent-updater, winagt-build and winagt-build-modules
The triggering job is setting the value of CIPARAM_PATH_HASH and can
thereby check for existing builds.

This does not require setting a custom build property nor evaluating
anything else then the Jenkins job parameters as it is done now with
other parameters.

The CUSTOM_GIT_REF is added to the no-check section of ci-artifacts
request. The job is by that still triggered with CUSTOM_GIT_REF being
set but not activly used to find the matching build. The matching build
is only found by CIPARAM_PATH_HASH

CMK-26429

Change-Id: I5c4d2e900af71dc4bf422d99d0e4680ad5c485f2