performance test: Allow baseline validation against old jobs
Instead of only comparing against the 4 previous jobs (=weekly baseline),
we also need to compare against jobs that are a month or even a year old.
To allow this, we add a --baseline-offset argument, which specifies the
number of days inbetween the baseline jobs and the current job. If this
is set to 30, 4 jobs will be fetched that are 30-33 days in the past.

CMK-25711

Change-Id: Ia87917aae13bf524feed2ec88f306637e8163579