18138 Fix select field in Chrome version 142+
A recent update of Chrome to version 142 changed the way select boxes
are being rendered.

This broke the styling of various multiple-select form fields like the
"Permission" selection when editing a host or when selecting files to
create an MKP from.

With this werk, the appearance of these select fields have been restored
to what they looked like before version 142.

CMK-27155

Change-Id: I2f70ee45dcbb07cf3db48f98d5ceb530af8dd876