Properly adapt integration tests to disabled EC in SaaS
The event console is disabled in SaaS environments, so some integration
tests need to be adapted accordingly.

It is not a 100% clear why this hasn't caused any test failures before,
but the assumption is that the SaaS specific behavior in
`omd/packages/check_mk/MKEVENTD` did not work as intended in the test
environment before.

CMK-27131

Change-Id: I604dc3eab48d7f27c3ad6304c670504c8511085b