move plugin consistency tests
These tests require all plugins to be present.
Since we start moving them to a dedicated package,
we have to execute these tests in an environment
where all packages are present.

Currently the quickest solution is to move them
to the integration tests.

Midterm, we should rather have a dedicated test
suite where we don't need a site.
This is subject of CMK-27183.

Change-Id: Ifac25b2ce02de724062e72da355cb831fc6f7eac