feat(unified-search) adjust provider search and rm click listener
As it is possible to inject a proper unified search from the other main
menu items it is more userfriendly not to dirctly head to the unified
search propup when clicking on to the search bar. Intead the injected
searchbar can be clicked to focus and the user is only headed to the
unified search popup when start typing.

Change-Id: I620ff61cbe37be509c9d2f8d1b2f2bc68134e6bb