gui(bi): make site and host mandatory vars
The underlying `bi_assumptions` dictionary expects to receive a tuple of
strings of length 2 or 3. However, with the current parsing, we could
receive `None` for site and host. This change explicitly states what the
get assumption by key receives and fixes the request variable parser.

Note, that this change changes behavior because the page will fail when
the `host` or `site` is not provide in the request.

Change-Id: I0206e88da956f2b25ae6a204b6272b77bfe4ded0