feat(slidein) add possiblity to teleport to content_area
For the removal of the iframe it will be necessary to be able to
teleport the slideout to another element than the body, as the ux team
wants the slideout not overlaying the navigation and sidebar.

Widt this change the posiblity is added to do this, but the current
behavior still is the default.

As soon as the main iframe is removed, the default has to be adjusted to
the main content element.

Change-Id: If3105b881882b5fe1a8ef520f4e48fa15e90a43a