gui(visuals): use request from ctx
This is an effort to reduce the access to the global `request` object.
We are now passing the request into the page handlers. This change
alters how handlers access the request, but is merely a structural
change.

In order to get the request down to the `Visual` class implementations,
we need to expand the `add_visual_handler` method to receive a request
parameter.

CMK-24397

Change-Id: I1a344ac615a831e21695f39ec44b782bfacd5cac