18344 FIX mssql_connections: Fix bug in cluster aggregation
When a clustered service for this plugin was configured, it was incorrectly returning results.
This result reported zero connections for nodes on which the service was not discovered.
This led to clustered services reporting results from multiple nodes, when only results from one host were available.

With this change the plugin does not report a result from nodes without that service.
The clustered service only shows the result from the nodes where the service is actually present.

CMK-26659
Change-Id: I21535386675a0332bd8908478aa6871c2583076c