Make v1 AgentConfig compatible with disk model
This unblocks upgrading to a version which does not use the legacy
converter anymore for the v1 AgentConfig. There, validation would
break as the disk model includes the "cmk-match-type" key which isn't
present in the form spec.

This key is required until all agent rulesets are using the v1 API.
Then, we can remove the special handling in the "boil_down_agent_rules"
function in `cmk/base/configlib/nonfree/pro/bakery.py`.

CMK-27243

Change-Id: Ib9e2f52623a2063dd75ed533150d5f33706d9f1e