tests/Makefile: Ensure that no empty paths are resolved
We should not be passing relative paths to realpath, at least not
without also providing the -m option, since otherwise the paths may
resolve to an empty string if one of the components does not exist.

From the man page of the realpath utility:
"Print the resolved absolute file name; all but the last component
must exist"

So for better readability, we should just pass only one component.

This will, for example, avoid running all tests when you really
just wanted to run the tests in a specific subfolder.

CMK-27327

Change-Id: I712b862d60e7c81a75bc1f7352c3825bf751ae8f