omd: inline unnecessary helpers
The problem described most likely refers to the fact that python keeps a
separate copy of the environment variables inside `os.environ`, which
differs from the standard one familiar to C programmers. This is
well-documented, for instance in the docs for `os.putenv` and does not
justify wrapping these calls.

The helpers couple functionality, that belongs to different modules.
Thus, we remove them.

CMK-27297

Change-Id: Ic8e00f11d9ce28321ce156f5a5b064c44f9caa02