cisco_meraki: move config into dedicated module
Added a nice builder method for getting the config from an args
namespace. I've also added some TODOs to inspect in the future.
Depending on the concrete Dashboard implementation makes this code
harder to test and to build config fixtures for other tests. Also, the
proxy that the Meraki SDK receives does not accept None as input, so
this would also need to be investigated.

Change-Id: I8b448c35990bcf5d1f0e041e0882452187f99b3d