Make implicit casting of cached data explicit
This allows you to act on a possible type change. Before, if stale data
was persisted in the *storage, it would cast it implicitly letting you
think it's certain that the data is of that type.

Add a no-op to the existing call sites, feel free to actually parse the
data instead.

Change-Id: I6d4fe60013d62710dffb86c91f2ad4508dcd5833