Upgrade pymssql to resolve compilation error.
Upgrade pymssql to include this change:

    https://github.com/pymssql/pymssql/pull/977

Resolves the following error:

    src/pymssql/_mssql.pyx:671:106: Unclosed string literal

This was a long standing bug in this package, but didn't cause
compilation problems with earlier versions of cython.

N.B. This version no longer bundles FreeTDS so we are no longer tied to the
same version.

This also drops our fork of pymssql, since the newer version compiles
with cython correctly.

https://github.com/TimotheusBachinger/pymssql@2a02c839c59c2b87e2acdfad8bc0187b2b6c4fe9

Change-Id: I8d564a4330ac2f47c235cb60854c8ffe7cd1dbe2