18710 FIX saml: editing existing configuration
Prior to this werk, saml2 configurations that had the idp metadata uploaded
as a file required the file to be re-uploaded in order to save the configuration
during editing.

This werk now displays the file content when inspecting the configuration and no longer
requires re-uploading the file. As a result, users can edit saml2 configurations
without the need to re-upload the idp metadata file each time they save. This werk also
includes an update action which adjusts the existing saml2 configurations.

SUP-26092

Change-Id: I6383aed4eb87e7826aecda5a4d1a8640bb73ffd0