17984 SEC Path-Traversal in report scheduler
revious to this Werk it was possible that an authenticated user could perform Path-traversal
attacks against the site's local file directory by use of the report scheduler. This issue was made
possible due to insufficient escaping of macros which could allow an attacker to make use of a
generated `.mk` file to overide existing `.mk` files.

Performing such an action could allow an attacker to break a site's configurations however as an
attacker cannot break out of the predefined fields within the generated `.mk` file, this can only
be used to DoS / break an affected site.

Change-Id: Ia4e236556e0918b7a3010143cc2c45db3eb79c2c