cisco_meraki: simplify caching for organizations
In the inbound MKP, Thomas consolidates this cache into one. The cache
doesn't concern itself with the org ids passed, but instead filters by
them before returning the organizations. This has the benefit of
reducing the complexity of the cache structure and will likely not
result in a performance penalty as the number of organizations shouldn't
be very large.

Change-Id: Ia4d75991fc3b1a4c8d628bd875dfe257f4043ed3