cisco_meraki: use internal meraki client for orgs
This acts as a wrapper to the meraki sdk dependency. This has the
advantage of being easier to test in the future and helps to return back
stronger types for our agent to process, i.e. `Organisation`.

Will want to follow up this change by tackling the other Cisco Meraki
resources that we are relying on like devices and licenses.

Change-Id: I690f4911933e10c7cc078b9725b696901727910b