cisco_meraki: migrate other sdk calls to clients
The main goal here is to make each of these client wrappers testable and
so that they return strong type schemas. We should be able to use these
same schemas on the sections on the plugin side to make it more type
compliant.

The individual schemas were generated using a tool [1], and the source
JSON specs can be found with their respective docstring links.

[1] https://json2pyi.pages.dev/#TypedDictClass

Change-Id: Ib235cf7546b3e7b165cd90bfdaae177f1e47185f