omd: remove --confirm-requires-root
It is clear (both from internal discussions and discussions with a user) that this option is confusing. That is unsurprising, since this option does nothing in the 2.5 and is only available in case we need an Apache update. I believe that there are better alternatives available to enforce a root update:

* Perform the privileged steps during the .deb/.rpm install.
* Require users to call `sudo omd update`.
* Install a binary with SUID bits for purposes of the update.

Change-Id: I4dd318b2fa910d45229d2cf5d9bd8cb12d57cd26