18909 FIX LDAP: Sync after create user on login should not remove users
When a user is created on first login, a binding is set for
that user but the bind was not being reset to the default
binding before running sync.  This could result in the loss
of existing users during the sync process since this user bind
may not have had permissions to access all users. This
behaviour has been adjusted and the binding is now reset as
soon as the user binding has been verified.

SUP-25379
SUP-25645

Change-Id: I02091578a4bd5d706921fd6920ab97d6f8c96c80