19007 Block Pro 2.4 to Pro 2.5 unlicensed update
Checkmk Pro 2.5 (formerly Enterprise) requires a valid license to operate.

If your Checkmk Pro 2.4 (formerly Enterprise) system is unlicensed or the license has expired, updates to 2.5 will be blocked.
This prevents licensed users from being downgraded to the trial or free version.

When an update is blocked, users will see an error message explaining the issue, and the system prevents the update from proceeding, keeping the system in a clean state as it was before the update started.

To proceed with the update, install a valid, non-expired license and then restart the update process.

For more information and detailed instructions please visit https://checkmk.atlassian.net/wiki/x/GIACIg

SAASDEV-5734

Change-Id: I455c437fafc564422e250eb36491d1820806d7eb