cisco_meraki: group dependencies into run context
A clean break from the former tests was needed to perform this change.
The new tests focus more on the `run` command and the associated agent
output. We're also now making use of polyfactory to generate data for
our fake SDK as opposed to explicitly defining the payload. Because of
this, the tests now define certain expected behaviors like there should
be less output if when filtering on section or org, but not asserting on
it's content.

Change-Id: I8d09c5dbfdd2ea08ac14cee35f558c9bb2e0e0e6