cisco_meraki: make clients callables and rename
It is more natural to have one client and then call methods on it that
reflect what is being returned. This should be fine as we are planning
to cache each resource individually, so no clients should contain more
than one public method.

Note, alternatively we could have made this a closure that returns a
callable, but I think it's clearer as a class and has the added bonus of
adding helper methods for the main call.

Change-Id: Ic6c2d4b2f39a817953249ef9f3e860bb345c4310