Speed up startup of special agents using `clickhouse-connect`
`clickhouse-connect` tries to import `numpy` but is fine with not finding it.
This import consumes significant CPU time. Since we don't need any functionality
requiring `numpy`, avoiding this import results in an immediate speed up.

This is of course just a temporary workaround. We are trying to solve this
upstream:
https://github.com/ClickHouse/clickhouse-connect/issues/589

Change-Id: I4849d760ef25dae93893c33c61fa43a338858815