cisco_meraki: decorate client query with cache
There is currently an effort to get rid of the `DataCache` helper.
Plugins are no longer allowed to rely on it. Therefore, internally we'll
support our own ttl cache decorator factory. This will then be used to
cache the other resources as well as incoming resources from the mkp.

Unfortunately, we can't make use of Python's semantic sugar by
decorating the callables with `@`. This is because we will need to
provide time to live (TTL) values at runtime.

Change-Id: If50f74b0f77bff34a72bd2953428413fc1a312b6