testlib.site: Add workaround for CMK-27517 and CMK-27518
We can not call openapi.sites.login() against a community edition
remote site with disabled replication, since neither the URL will
be set; nor the login endpoint works as expected.

As a temporary workaround, we just do not try to login after
creating the remote site connection until the issues are resolved.

Change-Id: Ie3c273d9acacbc581fcd96608deeff60d5ba32e0