gerrit: write out sections individually
This is a general refactor to improve the state of the gerrit plugin.
The primary goal is to not gather the sections and then write them out,
but instead to write them individually.

This allows us to use stronger schema types to provide to the section
writer helper. There's also an added bonus of being able to use these
types in the check logic. I will tackle this in a follow up.

I also needed to adapt existing tests to fit the new implementaton. But
I also took the liberty to improve the existing tests.

Note, a few mypy ignore flags were added because we are using the
underlying `get_data` from the `DataCache` instead of the unnecessary
`get_sections` method. These will be removed when the cache
implementation is swapped out.

Change-Id: Iff450118e76b61264c336024b4669b683e7d8307